[update] レイアウトの修正とディレクトリ名を直した
