+++ front/features/auth/components/sign-in-form.tsx
@@ -1,20 +1,34 @@
 "use client";
 
 import { useState } from "react";
-import { createUserWithEmailAndPassword } from "firebase/auth";
+import { signIn } from "next-auth/react";
+import { auth } from "@/libs/firebase/client";
+import { signInWithEmailAndPassword } from "firebase/auth";
 import Image from "next/image";
 import { Link, TextField } from "@mui/material";
-import { auth } from "@/libs/firebase/client";
 
-export const SignUpForm = () => {
+export const SignInForm = () => {
   const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
 
-  const handleSignUp = async () => {
+  const handleSignIn = async () => {
+    if (!email || !password) return;
+
     try {
-      const password = Math.random().toString(8).slice(2, 10);
-      const userCredential = await createUserWithEmailAndPassword(auth, email, password);
+      const credentials = await signInWithEmailAndPassword(auth, email, password);
+      const token = await credentials.user.getIdToken();
+      const result = await signIn("credentials", {
+        idToken: token,
+        redirect: false,
+      });
+
+      if (result?.error) {
+        console.error("認証エラー:", result.error);
+        return;
+      }
+
+      // 成功時の処理
       window.location.href = "/";
-      console.log(userCredential);
     } catch (error) {
       console.error(error);
     }
@@ -23,15 +37,29 @@ export const SignUpForm = () => {
   return (
     <>
       <div className="flex">
+        <div className="flex-1 relative h-full">
+          <Image
+            src="/login_page.png"
+            alt="Login page image"
+            className="object-cover w-full h-full"
+            width={864}
+            height={1000}
+          />
+        </div>
         <div className="flex-1 flex items-center justify-center">
           <div className="max-w-md p-8">
             <h1
               style={{ fontSize: "40px" }}
               className="font-bold leading-[40px] tracking-[0.21666669845581055px] text-center font-['October_Devanagari'] mb-2"
             >
-              アカウントの作成
+              おかえりなさい！
             </h1>
-
+            <p
+              style={{ fontSize: "14px" }}
+              className="font-bold leading-[14px] tracking-[0.21666669845581055px] text-center font-['October_Devanagari'] mb-8"
+            >
+              今日も素敵な写真をいっぱい投稿しましょう
+            </p>
             <div className="flex flex-col gap-6 w-[260px] mx-auto">
               <TextField
                 fullWidth
@@ -39,7 +67,6 @@ export const SignUpForm = () => {
                 value={email}
                 onChange={(e) => setEmail(e.target.value)}
                 variant="outlined"
-                style={{ marginTop: "48px" }}
                 InputLabelProps={{
                   shrink: true,
                   sx: {
@@ -59,14 +86,13 @@ export const SignUpForm = () => {
                   },
                 }}
               />
-              {/* <TextField
+              <TextField
                 fullWidth
                 type="password"
                 label="パスワード"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
                 variant="outlined"
-                style={{ marginTop: "16px" }}
                 InputLabelProps={{
                   shrink: true,
                   sx: {
@@ -85,27 +111,27 @@ export const SignUpForm = () => {
                     },
                   },
                 }}
-              /> */}
+              />
               <button
-                onClick={handleSignUp}
+                onClick={handleSignIn}
                 style={{
                   height: "56px",
-                  marginTop: "32px",
+                  marginTop: "16px",
                 }}
                 className="bg-[#69BEEF] text-white rounded-md hover:bg-[#5CAADB]"
               >
-                作成する
+                ログイン
               </button>
-              <p
-                style={{ marginTop: "16px" }}
-                className="font-['October_Devanagari'] text-[10px] font-[400] leading-[16px] text-center"
-              >
-                既にアカウントをお持ちですか？{" "}
-                <Link href="/sign-in" className="underline">
-                  ログイン
+              <p className="font-['October_Devanagari'] text-[10px] font-[400] leading-[16px] text-center">
+                アカウントをお持ちではありませんか？{" "}
+                <Link href="/sign-up" className="underline">
+                  アカウント作成
                 </Link>
               </p>
-              <p style={{ marginTop: "32px", fontSize: "10px" }} className="text-center text-[#000000] mt-[30px]">
+              <p
+                className="text-center text-[#000000]"
+                style={{ marginTop: "16px", fontSize: "10px" }}
+              >
                 ---------------------------または---------------------------
               </p>
               <button
@@ -115,7 +141,6 @@ export const SignUpForm = () => {
                   display: "flex",
                   alignItems: "center",
                   justifyContent: "center",
-                  marginTop: "16px",
                 }}
                 className="text-[#000000] rounded-md hover:bg-gray-200"
               >
@@ -124,15 +149,6 @@ export const SignUpForm = () => {
               </button>
             </div>
           </div>
-        </div>
-        <div className="flex-1 relative h-full">
-          <Image
-            src="/signup-icon.png"
-            alt="Login page image"
-            className="object-cover w-full h-full"
-            width={864}
-            height={800}
-          />
         </div>
       </div>
     </>
