+++ front/features/auth/components/sign-up-form.tsx
@@ -1,34 +1,19 @@
 "use client";
 
 import { useState } from "react";
-import { signIn } from "next-auth/react";
-import { auth } from "@/libs/firebase/client";
-import { signInWithEmailAndPassword } from "firebase/auth";
+import { createUserWithEmailAndPassword } from "firebase/auth";
 import Image from "next/image";
 import { Link, TextField } from "@mui/material";
+import { auth } from "@/libs/firebase/client";
 
-export const SignInForm = () => {
+export const SignUpForm = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-
-  const handleSignIn = async () => {
-    if (!email || !password) return;
-
+  const handleSignUp = async () => {
     try {
-      const credentials = await signInWithEmailAndPassword(auth, email, password);
-      const token = await credentials.user.getIdToken();
-      const result = await signIn("credentials", {
-        idToken: token,
-        redirect: false,
-      });
-
-      if (result?.error) {
-        console.error("認証エラー:", result.error);
-        return;
-      }
-
-      // 成功時の処理
+      const userCredential = await createUserWithEmailAndPassword(auth, email, password);
       window.location.href = "/";
+      console.log(userCredential);
     } catch (error) {
       console.error(error);
     }
@@ -37,29 +22,15 @@ export const SignInForm = () => {
   return (
     <>
       <div className="flex">
-        <div className="flex-1 relative h-full">
-          <Image
-            src="/login_page.png"
-            alt="Login page image"
-            className="object-cover w-full h-full"
-            width={864}
-            height={1000}
-          />
-        </div>
         <div className="flex-1 flex items-center justify-center">
           <div className="max-w-md p-8">
             <h1
               style={{ fontSize: "40px" }}
               className="font-bold leading-[40px] tracking-[0.21666669845581055px] text-center font-['October_Devanagari'] mb-2"
             >
-              おかえりなさい！
+              アカウントの作成
             </h1>
-            <p
-              style={{ fontSize: "14px" }}
-              className="font-bold leading-[14px] tracking-[0.21666669845581055px] text-center font-['October_Devanagari'] mb-8"
-            >
-              今日も素敵な写真をいっぱい投稿しましょう
-            </p>
+
             <div className="flex flex-col gap-6 w-[260px] mx-auto">
               <TextField
                 fullWidth
@@ -67,6 +38,7 @@ export const SignInForm = () => {
                 value={email}
                 onChange={(e) => setEmail(e.target.value)}
                 variant="outlined"
+                style={{ marginTop: "48px" }}
                 InputLabelProps={{
                   shrink: true,
                   sx: {
@@ -113,23 +85,25 @@ export const SignInForm = () => {
                 }}
               />
               <button
-                onClick={handleSignIn}
+                onClick={handleSignUp}
                 style={{
                   height: "56px",
+                  marginTop: "8px",
                 }}
                 className="bg-[#69BEEF] text-white rounded-md hover:bg-[#5CAADB]"
               >
-                ログイン
+                作成する
               </button>
-              <p
-                className="font-['October_Devanagari'] text-[10px] font-[400] leading-[16px] text-center"
-              >
-                アカウントをお持ちではありませんか？{" "}
-                <Link href="/sign-up" className="underline">
-                  アカウント作成
+              <p className="font-['October_Devanagari'] text-[10px] font-[400] leading-[16px] text-center">
+                既にアカウントをお持ちですか？{" "}
+                <Link href="/sign-in" className="underline">
+                  ログイン
                 </Link>
               </p>
-              <p className="text-center text-[#000000]" style={{ marginTop: "32px", fontSize: "10px" }}>
+              <p
+                style={{ marginTop: "16px", fontSize: "10px" }}
+                className="text-center text-[#000000] mt-[30px]"
+              >
                 ---------------------------または---------------------------
               </p>
               <button
@@ -147,6 +121,15 @@ export const SignInForm = () => {
               </button>
             </div>
           </div>
+        </div>
+        <div className="flex-1 relative h-full">
+          <Image
+            src="/signup-icon.png"
+            alt="Login page image"
+            className="object-cover w-full h-full"
+            width={864}
+            height={800}
+          />
         </div>
       </div>
     </>
